requirements.txt
jsons
requests
pillow
FlightRadarAPI
